+++ tests/code/qubits/pauli.rs
