+++ tests/code/qubits/pauligroup.rs
