+++ tests/lib.rs
@@ -1,4 +1,4 @@
-mod codes;
+mod code;
 mod map;
 mod math;
 mod serialize;
